Applied many revisions recommended by the code analyzer.
Uncovered an issue where the interval was not being used within the animation routine, effectively treating each interval as if it were one second.  Modified Game ViewModel to use the interval as intended and changed the minimum and maximum speeds to accommodate the smaller, correct time slices.

+++ app/src/main/java/com/twilightcitizen/whack_a_pede/utilities/TimeUtil.java
@@ -26,10 +26,10 @@ public class TimeUtil {
     }
 
     // Convert milliseconds into a fractional interval of seconds.
-    public static Double millisToIntervalOfSeconds( long millis ) { return millis / 1000d; }
+    public static float millisToIntervalOfSeconds( long millis ) { return millis / 1000.0f; }
 
     // Convert seconds to milliseconds.
-    public static long secondsToMillis( int seconds ) { return seconds * 1000; }
+    public static long secondsToMillis( int seconds ) { return seconds * 1000L; }
 
     // Convert milliseconds to whole seconds.
     public static int millisToSeconds( long millis ) { return (int) millis / 60; }
